backend/requirements.txt
fastapi
uvicorn
yfinance
newspaper3k
googlenews
pydantic
python-dotenv
google-generativeai
dateparser
duckduckgo-search
requests
beautifulsoup4
googlesearch-python
